requirements.txt
Flask==2.3.2
pyOpenSSL==24.0.0
